+++ LuckyNumbers.API/Dtos/ResultLottoDto.cs
@@ -13,7 +13,7 @@ namespace LuckyNumbers.API.Dtos
         public int goal6Numbers { get; set; }
         public int totalCostBets { get; set; }
         public int totalEarnMoney { get; set; }
-        public int totalEarnExp { get; set; }
+        public double totalEarnExp { get; set; }
         public List<LottoNumbersDto> betsWithGoal3Numbers { get; set; } = new List<LottoNumbersDto>();
         public List<LottoNumbersDto> betsWithGoal4Numbers { get; set; } = new List<LottoNumbersDto>();
         public List<LottoNumbersDto> betsWithGoal5Numbers { get; set; } = new List<LottoNumbersDto>();
